feat(runtime): widen portable profile capacity
Carry target-profile capacity as u64 through the semantic requirement and
preflight boundary while preserving frozen effect IR v3 and MBPF v3 u32 fields.
Reject wider envelopes before cache/native identity, keep direct narrowing
checked, and restore the immutable 2026.3 byte fixture to its own profile.

+++ src/runtime/tiered-execution/adapter-outbound/cache/main.rs
@@ -508,9 +508,7 @@ impl RegionEffectIdentity {
         digest: BucketDigestFunction,
     ) -> Result<Self, NativeIdentityError> {
         let required_memory_words = program.required_memory_words();
-        if required_memory_words
-            > u64::from(program.profile_requirement.memory_words)
-        {
+        if required_memory_words > program.profile_requirement.memory_words {
             return Err(NativeIdentityError::ProfileCapacity);
         }
         let canonical = program.canonical_bytes()?;
